mess about with some structure to help w/ deref
